refactor: split bulk upload tests into focused files
- Split test_bulk_upload.py into 4 focused files:
  - test_bulk_upload_endpoints.py: API endpoint tests
  - test_bulk_upload_manager.py: Manager functionality tests
  - test_bulk_upload_patterns.py: Filename pattern parsing tests
  - test_bulk_upload_utils.py: Utility function tests
- Fixed all flake8 errors (unused imports, variables, missing newlines)
- Maintained all original test functionality and coverage
- Improved maintainability and parallel test execution
